security enhancements

+++ api/src/main/java/com/epam/esm/util/ModelAssembler.java
@@ -5,10 +5,15 @@ import com.epam.esm.controller.OrderRestController;
 import com.epam.esm.controller.TagRestController;
 import com.epam.esm.controller.UserRestController;
 import com.epam.esm.dto.CertificateDTO;
+import com.epam.esm.dto.CertificateListDTO;
 import com.epam.esm.dto.CertificatePatchDTO;
 import com.epam.esm.dto.OrderDTO;
+import com.epam.esm.dto.OrderListDTO;
 import com.epam.esm.dto.TagDTO;
+import com.epam.esm.dto.TagListDTO;
 import com.epam.esm.dto.UserDTO;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
 
 import javax.servlet.http.HttpServletResponse;
 
@@ -24,9 +29,18 @@ public class ModelAssembler {
   }
 
   public static void addTagLinks(TagDTO tagDTO, HttpServletResponse resp) {
-    tagDTO.add(
-        linkTo(methodOn(TagRestController.class).getById(tagDTO.getId(), resp)).withSelfRel());
-    tagDTO.add(linkTo(methodOn(TagRestController.class).delete(tagDTO.getId())).withRel("delete"));
+    addTagSelfLink(tagDTO, resp);
+    if (appUserIsAdmin()) {
+      tagDTO.add(
+          linkTo(methodOn(TagRestController.class).delete(tagDTO.getId())).withRel("delete"));
+    }
+  }
+
+  public static void addTagListLinks(TagListDTO tagListDTO) {
+    if (appUserIsAdmin()) {
+      tagListDTO.add(
+          linkTo(methodOn(TagRestController.class).create(new TagDTO())).withRel("create"));
+    }
   }
 
   public static void addCertificateSelfLink(
@@ -39,24 +53,32 @@ public class ModelAssembler {
 
   public static void addCertificateLinks(CertificateDTO certificateDTO, HttpServletResponse resp) {
     addCertificateSelfLink(certificateDTO, resp);
-    certificateDTO.add(
-        linkTo(
-                methodOn(CertificateRestController.class)
-                    .update(certificateDTO.getId(), new CertificateDTO()))
-            .withRel("update"));
-    certificateDTO.add(
-        linkTo(
-                methodOn(CertificateRestController.class)
-                    .patch(certificateDTO.getId(), new CertificatePatchDTO()))
-            .withRel("patch"));
-    if (!certificateDTO.isDeleted()) {
+    if (appUserIsAdmin()) {
       certificateDTO.add(
-          linkTo(methodOn(CertificateRestController.class).delete(certificateDTO.getId()))
-              .withRel("delete"));
+          linkTo(methodOn(CertificateRestController.class)
+                      .update(certificateDTO.getId(), new CertificateDTO()))
+              .withRel("update"));
+      certificateDTO.add(
+          linkTo(methodOn(CertificateRestController.class)
+                      .patch(certificateDTO.getId(), new CertificatePatchDTO()))
+              .withRel("patch"));
+      if (!certificateDTO.isDeleted()) {
+        certificateDTO.add(
+            linkTo(methodOn(CertificateRestController.class).delete(certificateDTO.getId()))
+                .withRel("delete"));
+      }
     }
     certificateDTO.add(
         linkTo(methodOn(CertificateRestController.class).getAll(null, null, null, "1", "10", resp))
             .withRel("getAll"));
+  }
+
+  public static void addCertificateListLinks(CertificateListDTO certificateListDTO) {
+    if (appUserIsAdmin()) {
+      certificateListDTO.add(
+          linkTo(methodOn(CertificateRestController.class).create(new CertificateDTO()))
+              .withRel("create"));
+    }
   }
 
   public static void addOrderSelfLink(OrderDTO orderDTO, HttpServletResponse resp) {
@@ -80,6 +102,14 @@ public class ModelAssembler {
             .withRel("getAll"));
   }
 
+  public static void addOrderListLinks(OrderListDTO orderListDTO, HttpServletResponse resp) {
+    Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    int appUserId = principal instanceof Integer ? (Integer) principal : 1;
+    orderListDTO.add(
+        linkTo(methodOn(UserRestController.class).createOrder(appUserId, new OrderDTO(), resp))
+            .withRel("create"));
+  }
+
   public static void addUserSelfLink(UserDTO userDTO, HttpServletResponse resp) {
     userDTO.add(
         linkTo(methodOn(UserRestController.class).getUserById(userDTO.getId(), resp))
@@ -88,12 +118,14 @@ public class ModelAssembler {
 
   public static void addUserLinks(UserDTO userDTO, HttpServletResponse resp) {
     addUserSelfLink(userDTO, resp);
-    if (!userDTO.isDeleted()) {
+    if (appUserIsAdmin()) {
+      if (!userDTO.isDeleted()) {
+        userDTO.add(
+            linkTo(methodOn(UserRestController.class).deleteUser(userDTO.getId())).withRel("delete"));
+      }
       userDTO.add(
-          linkTo(methodOn(UserRestController.class).deleteUser(userDTO.getId())).withRel("delete"));
+          linkTo(methodOn(UserRestController.class).getAllUsers("1", "10", resp)).withRel("getAll"));
     }
-    userDTO.add(
-        linkTo(methodOn(UserRestController.class).getAllUsers("1", "10", resp)).withRel("getAll"));
   }
 
   public static void addUsersOrderSelfLink(
@@ -117,5 +149,14 @@ public class ModelAssembler {
     orderDTO.add(
         linkTo(methodOn(UserRestController.class).getUserOrders(userId, "1", "10", resp))
             .withRel("getAll"));
+  }
+
+  private static boolean appUserIsAdmin() {
+    String appUserRole =
+        SecurityContextHolder.getContext().getAuthentication().getAuthorities().stream()
+            .findFirst()
+            .map(GrantedAuthority::getAuthority)
+            .orElse("ROLE_GUEST");
+    return appUserRole.equals("ROLE_ADMIN");
   }
 }
